requirements.txt
opencv-python
numpy
tensorflow
matplotlib
pillow
